requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
PyYAML==6.0.3
sqlparse==0.5.5
